bugfix; broker drain duration ought to be a float. (#34)
What:
- bugfix; broker drain duration ought to be a float.

Why:
- that is what is been implemented upstream: https://github.com/komuw/wiji/pull/34

+++ wijisqs/broker.py
@@ -767,7 +767,7 @@ class SqsBroker(wiji.broker.BaseBroker):
             )
             self.logger.log(logging.DEBUG, response)
 
-    async def shutdown(self, queue_name: str, duration: int) -> None:
+    async def shutdown(self, queue_name: str, duration: float) -> None:
         """
         when this method is called:
           1. wijisqs should halt all consumption from AWS sqs
